+++ CHANGELOG.md
@@ -191,4 +191,6 @@ Encounter typescipt bug, VSC 1.94.2, Node.js 20.16.0, Npm 10.8.2
 
 ## 0.12.2024120501
 
+- Fix current repo not sync with selected repo
+
 - Watch git root file system change